Package files in a more organized way
