--- data/tasks.txt
@@ -1,12 +1,0 @@
-T | 0 | 1111111111111
-T | 0 | 22222222222222
-T | 0 | 33333333333
-T | 0 | ninininin
-T | 0 | dhiewhdi
-T | 0 | sdhiuewg
-T | 0 | doiweh
-T | 0 | whdioew
-T | 0 | dwuhfiuwegtufi
-T | 0 | udshiuw
-T | 0 | suedhfui
-T | 0 | 222
