+++ src/main/java/jiayou/view/DialogBox.java
@@ -1,4 +1,4 @@
-package jiayou;
+package jiayou.view;
 
 import java.io.IOException;
 import java.util.Collections;
