+++ src/main/java/jiayou/view/Main.java
@@ -1,4 +1,4 @@
-package jiayou;
+package jiayou.view;
 
 import java.io.IOException;
 
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import jiayou.function.Jiayou;
 
 /**
  * A GUI for Jiayou using FXML.
