+++ src/main/java/jiayou/view/MainWindow.java
@@ -1,4 +1,4 @@
-package jiayou;
+package jiayou.view;
 
 import javafx.animation.PauseTransition;
 import javafx.fxml.FXML;
@@ -8,6 +8,8 @@ import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
 import javafx.util.Duration;
+import jiayou.function.Jiayou;
+
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
  */
