+++ src/main/resources/view/MainWindow.fxml
@@ -6,13 +6,17 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="600.0" style="-fx-background-color: #add8e6;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="jiayou.MainWindow">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="600.0" style="-fx-background-color: #feffd1;" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="jiayou.view.MainWindow">
     <children>
-        <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="530" AnchorPane.bottomAnchor="1.0" />
-        <Button fx:id="sendButton" layoutX="530.0" layoutY="558.0" mnemonicParsing="false" onAction="#handleUserInput" prefHeight="41.0" prefWidth="76.0" style="-fx-cursor: hand;" text="Send" />
+        <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="525"  AnchorPane.bottomAnchor="1.0"
+        style="-fx-background-radius: 10px; -fx-border-radius: 10px; -fx-padding: 5px;">
+    </TextField>
+        <Button fx:id="sendButton" layoutX="525.0" layoutY="558.0" mnemonicParsing="false" onAction="#handleUserInput" prefHeight="41.0" prefWidth="76.0" text="Send"
+                style="-fx-font-size: 15; -fx-cursor: hand; -fx-text-fill: black; -fx-font-weight: bold; -fx-border-radius: 10px; -fx-background-radius: 15px; -fx-padding: 5px;">
+        </Button>
         <ScrollPane fx:id="scrollPane" hbarPolicy="NEVER" hvalue="1.0" prefHeight="557.0" prefWidth="605.0" style="-fx-background-color: #feffd1;" vvalue="1.0">
             <content>
-                <VBox fx:id="dialogContainer" prefHeight="-1" prefWidth="588.0" style="-fx-background-color: #feffd1;" AnchorPane.rightAnchor="0.0" />
+                <VBox fx:id="dialogContainer"  minHeight="${scrollPane.prefHeight}" prefWidth="588.0" style="-fx-background-color: #feffd1;" AnchorPane.rightAnchor="0.0" />
             </content>
         </ScrollPane>
     </children>
